+++ capytaine/bem/problems_and_results.py
@@ -17,7 +17,8 @@ LOG = logging.getLogger(__name__)
 
 _default_parameters = {'rho': 1000.0, 'g': 9.81, 'omega': 1.0,
                       'free_surface': 0.0, 'water_depth': np.inf,
-                      'wave_direction': 0.0}
+                       'wave_direction': 0.0, 'forward_speed': 0.0}
+
 
 
 class LinearPotentialFlowProblem:
@@ -44,6 +45,8 @@ class LinearPotentialFlowProblem:
         The angular wave number of the waves in rad/m
     wavelength: float, optional
         The wave length of the waves in m
+    forward_speed: float, optional
+        The speed of the body (in m/s, in the x direction, default: 0.0)
     rho: float, optional
         The density of water in kg/m3 (default: 1000.0)
     g: float, optional
@@ -57,24 +60,45 @@ class LinearPotentialFlowProblem:
                  free_surface=_default_parameters['free_surface'],
                  water_depth=None, sea_bottom=None,
                  omega=None, period=None, wavenumber=None, wavelength=None,
+                 forward_speed=_default_parameters['forward_speed'],
                  rho=_default_parameters['rho'],
                  g=_default_parameters['g'],
+                 wave_direction=_default_parameters['wave_direction'],
                  boundary_condition=None):
 
         self.body = body
         self.free_surface = float(free_surface)
         self.rho = float(rho)
         self.g = float(g)
+        self.forward_speed = float(forward_speed)
+        self.wave_direction = float(wave_direction)  # Required for (diffraction problem) and (radiation problems with forward speed).
 
         self.boundary_condition = boundary_condition
 
         self.water_depth = _get_water_depth(free_surface, water_depth, sea_bottom, default_water_depth=_default_parameters["water_depth"])
         self.omega, self.period, self.wavenumber, self.wavelength, self.provided_freq_type = \
-                self._get_frequencies(omega, period, wavenumber, wavelength)
+                self._get_frequencies(omega=omega, period=period, wavenumber=wavenumber, wavelength=wavelength)
 
         self._check_data()
 
-    def _get_frequencies(self, omega, period, wavenumber, wavelength):
+        if forward_speed != 0.0:
+            dopplered_omega = self.omega - self.wavenumber*self.forward_speed*np.cos(self.wave_direction)
+            self.encounter_omega, self.encounter_period, self.encounter_wavenumber, self.encounter_wavelength, _ = \
+                    self._get_frequencies(omega=abs(dopplered_omega))
+
+            if dopplered_omega >= 0.0:
+                self.encounter_wave_direction = self.wave_direction
+            else:
+                self.encounter_wave_direction = self.wave_direction + np.pi
+        else:
+            self.encounter_omega = self.omega
+            self.encounter_period = self.period
+            self.encounter_wavenumber = self.wavenumber
+            self.encounter_wavelength = self.wavelength
+            self.encounter_wave_direction = self.wave_direction
+
+
+    def _get_frequencies(self, *, omega=None, period=None, wavenumber=None, wavelength=None):
         frequency_data = dict(omega=omega, period=period, wavenumber=wavenumber, wavelength=wavelength)
         nb_provided_frequency_data = 4 - list(frequency_data.values()).count(None)
 
@@ -138,7 +162,13 @@ class LinearPotentialFlowProblem:
                 "Only z=0 and z=∞ are accepted values for the free surface position."
             )
 
+        if not (-2*np.pi-1e-3 <= self.wave_direction <= 2*np.pi+1e-3):
+            LOG.warning(f"The value {self.wave_direction} has been provided for the wave direction, and it does not look like an angle in radians. "
+                         "The wave direction in Capytaine is defined in radians and not in degrees, so the result might not be what you expect. "
+                         "If you were actually giving an angle in radians, use the modulo operator to give a value between -2π and 2π to disable this warning.")
+
         if self.free_surface == np.inf and self.water_depth != np.inf:
+
             raise NotImplementedError(
                 "Problems with a sea bottom but no free surface have not been implemented."
             )
@@ -146,10 +176,17 @@ class LinearPotentialFlowProblem:
         if self.water_depth < 0.0:
             raise ValueError("`water_depth` should be strictly positive (provided water depth: {self.water_depth}).")
 
-        if float(self.omega) in {0, np.inf} and self.water_depth != np.inf:
-            LOG.warning(
-                    f"Default Green function allows for {self.provided_freq_type}={float(self.__getattribute__(self.provided_freq_type))} only for infinite depth (provided water depth: {self.water_depth})."
-            )
+        if float(self.omega) in {0, np.inf}:
+            if self.water_depth != np.inf:
+                LOG.warning(
+                        f"Default Green function allows for {self.provided_freq_type}={float(self.__getattribute__(self.provided_freq_type))} only for infinite depth (provided water depth: {self.water_depth})."
+                        )
+
+            if self.forward_speed != 0.0:
+                raise NotImplementedError(
+                        f"omega={float(self.omega)} is only implemented without forward speed (provided forward speed: {self.forward_speed})."
+                )
+
 
         if self.body is not None:
             if ((isinstance(self.body.mesh, CollectionOfMeshes) and len(self.body.mesh) == 0)
@@ -183,9 +220,13 @@ class LinearPotentialFlowProblem:
         return {"body_name": self.body_name,
                 "water_depth": self.water_depth,
                 "omega": self.omega,
+                "encounter_omega": self.encounter_omega,
                 "period": self.period,
                 "wavelength": self.wavelength,
                 "wavenumber": self.wavenumber,
+                "forward_speed": self.forward_speed,
+                "wave_direction": self.wave_direction,
+                "encounter_wave_direction": self.encounter_wave_direction,
                 "rho": self.rho,
                 "g": self.g}
 
@@ -205,6 +246,10 @@ class LinearPotentialFlowProblem:
         parameters = [f"body={self.body.__short_str__() if self.body is not None else None}",
                       f"{self.provided_freq_type}={self.__getattribute__(self.provided_freq_type):.3f}",
                       f"water_depth={self.water_depth}"]
+
+        if not self.forward_speed == _default_parameters['forward_speed']:
+            parameters.append(f"forward_speed={self.forward_speed:.3f}")
+
         try:
             parameters.extend(self._str_other_attributes())
         except AttributeError:
@@ -239,7 +284,7 @@ class LinearPotentialFlowProblem:
     def _astuple(self):
         return (self.body, self.free_surface, self.water_depth,
                 self.omega, self.period, self.wavenumber, self.wavelength,
-                self.rho, self.g)
+                self.forward_speed, self.rho, self.g)
 
     def __eq__(self, other):
         if isinstance(other, LinearPotentialFlowProblem):
@@ -278,19 +323,14 @@ class DiffractionProblem(LinearPotentialFlowProblem):
                  free_surface=_default_parameters['free_surface'],
                  water_depth=None, sea_bottom=None,
                  omega=None, period=None, wavenumber=None, wavelength=None,
+                 forward_speed=_default_parameters['forward_speed'],
                  rho=_default_parameters['rho'],
                  g=_default_parameters['g'],
                  wave_direction=_default_parameters['wave_direction']):
 
-        self.wave_direction = float(wave_direction)
-
         super().__init__(body=body, free_surface=free_surface, water_depth=water_depth, sea_bottom=sea_bottom,
-                         omega=omega, period=period, wavenumber=wavenumber, wavelength=wavelength, rho=rho, g=g)
-
-        if not (-2*np.pi-1e-3 <= self.wave_direction <= 2*np.pi+1e-3):
-            LOG.warning(f"The value {self.wave_direction} has been provided for the wave direction, and it does not look like an angle in radians. "
-                         "The wave direction in Capytaine is defined in radians and not in degrees, so the result might not be what you expect. "
-                         "If you were actually giving an angle in radians, use the modulo operator to give a value between -2π and 2π to disable this warning.")
+                         omega=omega, period=period, wavenumber=wavenumber, wavelength=wavelength, wave_direction=wave_direction,
+                         forward_speed=forward_speed, rho=rho, g=g)
 
         if float(self.omega) in {0.0, np.inf}:
             raise NotImplementedError(f"DiffractionProblem does not support zero or infinite frequency.")
@@ -301,17 +341,11 @@ class DiffractionProblem(LinearPotentialFlowProblem):
                     airy_waves_velocity(self.body.mesh.faces_centers, self)
                     * self.body.mesh.faces_normals
             ).sum(axis=1)
+            # Note that even with forward speed, this is computed based on the
+            # frequency and not the encounter frequency.
 
             if len(self.body.dofs) == 0:
                 LOG.warning(f"The body {self.body.name} used in diffraction problem has no dofs!")
-
-    def _astuple(self):
-        return super()._astuple() + (self.wave_direction,)
-
-    def _asdict(self):
-        d = super()._asdict()
-        d["wave_direction"] = self.wave_direction
-        return d
 
     def _str_other_attributes(self):
         return [f"wave_direction={self.wave_direction:.3f}"]
@@ -331,6 +365,8 @@ class RadiationProblem(LinearPotentialFlowProblem):
                  free_surface=_default_parameters['free_surface'],
                  water_depth=None, sea_bottom=None,
                  omega=None, period=None, wavenumber=None, wavelength=None,
+                 forward_speed=_default_parameters['forward_speed'],
+                 wave_direction=_default_parameters['wave_direction'],
                  rho=_default_parameters['rho'],
                  g=_default_parameters['g'],
                  radiating_dof=None):
@@ -338,7 +374,8 @@ class RadiationProblem(LinearPotentialFlowProblem):
         self.radiating_dof = radiating_dof
 
         super().__init__(body=body, free_surface=free_surface, water_depth=water_depth, sea_bottom=sea_bottom,
-                         omega=omega, period=period, wavenumber=wavenumber, wavelength=wavelength, rho=rho, g=g)
+                         omega=omega, period=period, wavenumber=wavenumber, wavelength=wavelength,
+                         wave_direction=wave_direction, forward_speed=forward_speed, rho=rho, g=g)
 
         if self.body is not None:
 
@@ -355,7 +392,24 @@ class RadiationProblem(LinearPotentialFlowProblem):
                 raise ValueError("Unrecognized degree of freedom name.")
 
             dof = self.body.dofs[self.radiating_dof]
-            self.boundary_condition = -1j*self.omega * np.sum(dof * self.body.mesh.faces_normals, axis=1)
+
+            self.boundary_condition = -1j * self.encounter_omega * np.sum(dof * self.body.mesh.faces_normals, axis=1)
+
+            if self.forward_speed != 0.0:
+                if self.radiating_dof.lower() == "pitch":
+                    ddofdx_dot_n = np.array([nz for (nx, ny, nz) in self.body.mesh.faces_normals])
+                elif self.radiating_dof.lower() == "yaw":
+                    ddofdx_dot_n = np.array([-ny for (nx, ny, nz) in self.body.mesh.faces_normals])
+                elif self.radiating_dof.lower() in {"surge", "sway", "heave", "roll"}:
+                    ddofdx_dot_n = 0.0
+                else:
+                    raise NotImplementedError(
+                            "Radiation problem with forward speed is currently only implemented for a single rigid body.\n"
+                            "Only radiating dofs with name in {'Surge', 'Sway', 'Heave', 'Roll', 'Pitch', 'Yaw'} are supported.\n"
+                            f"Got instead `radiating_dof={self.radiating_dof}`"
+                            )
+                self.boundary_condition += self.forward_speed * ddofdx_dot_n
+
 
     def _astuple(self):
         return super()._astuple() + (self.radiating_dof,)
@@ -366,7 +420,10 @@ class RadiationProblem(LinearPotentialFlowProblem):
         return d
 
     def _str_other_attributes(self):
-        return [f"radiating_dof=\'{self.radiating_dof}\'"]
+        if self.forward_speed != 0.0:
+            return [f"wave_direction={self.wave_direction:.3f}, radiating_dof=\'{self.radiating_dof}\'"]
+        else:
+            return [f"radiating_dof=\'{self.radiating_dof}\'"]
 
     def _specific_rich_repr(self):
         yield "radiating_dof", self.radiating_dof
@@ -391,14 +448,21 @@ class LinearPotentialFlowResult:
         self.body               = self.problem.body
         self.free_surface       = self.problem.free_surface
         self.omega              = self.problem.omega
+        self.period             = self.problem.period
+        self.wavenumber         = self.problem.wavenumber
+        self.wavelength         = self.problem.wavelength
+        self.forward_speed      = self.problem.forward_speed
+        self.wave_direction     = self.problem.wave_direction
+        self.encounter_omega    = self.problem.encounter_omega
+        self.encounter_period   = self.problem.encounter_period
+        self.encounter_wavenumber = self.problem.encounter_wavenumber
+        self.encounter_wavelength = self.problem.encounter_wavelength
+        self.encounter_wave_direction = self.problem.encounter_wave_direction
         self.rho                = self.problem.rho
         self.g                  = self.problem.g
         self.boundary_condition = self.problem.boundary_condition
         self.water_depth        = self.problem.water_depth
         self.depth              = self.problem.water_depth
-        self.wavenumber         = self.problem.wavenumber
-        self.wavelength         = self.problem.wavelength
-        self.period             = self.problem.period
         self.provided_freq_type = self.problem.provided_freq_type
         self.body_name          = self.problem.body_name
         self.influenced_dofs    = self.problem.influenced_dofs
@@ -418,7 +482,6 @@ class DiffractionResult(LinearPotentialFlowResult):
 
     def __init__(self, problem, *args, **kwargs):
         super().__init__(problem, *args, **kwargs)
-        self.wave_direction = self.problem.wave_direction
 
     _str_other_attributes = DiffractionProblem._str_other_attributes
     _specific_rich_repr = DiffractionProblem._specific_rich_repr
@@ -445,11 +508,11 @@ class RadiationResult(LinearPotentialFlowResult):
 
     @property
     def added_mass(self):
-        return {dof: float(np.real(force)/(self.omega*self.omega)) for (dof, force) in self.forces.items()}
+        return {dof: float(np.real(force)/(self.encounter_omega*self.encounter_omega)) for (dof, force) in self.forces.items()}
 
     @property
     def radiation_damping(self):
-        return {dof: float(np.imag(force)/self.omega) for (dof, force) in self.forces.items()}
+        return {dof: float(np.imag(force)/self.encounter_omega) for (dof, force) in self.forces.items()}
 
     # Aliases for backward compatibility
     added_masses = added_mass