+++ capytaine/green_functions/libDelhommeau/src/matrices.f90
@@ -19,7 +19,7 @@ CONTAINS
       nb_quad_points, quad_points, quad_weights,      &
       wavenumber, depth,                              &
       coeffs,                                         &
-      tabulated_r_range, tabulated_z_range, tabulated_integrals, &
+      tabulation_method, tabulated_r_range, tabulated_z_range, tabulated_integrals, &
       NEXP, AMBDA, AR,                                &
       same_body, adjoint_double_layer,                &
       S, K)
@@ -44,6 +44,7 @@ CONTAINS
     REAL(KIND=PRE), DIMENSION(3)                         :: coeffs
 
     ! Tabulated data
+    INTEGER,                                  INTENT(IN) :: tabulation_method
     REAL(KIND=PRE), DIMENSION(:),             INTENT(IN) :: tabulated_r_range
     REAL(KIND=PRE), DIMENSION(:),             INTENT(IN) :: tabulated_z_range
     REAL(KIND=PRE), DIMENSION(:, :, :, :),    INTENT(IN) :: tabulated_integrals
@@ -188,7 +189,7 @@ CONTAINS
             centers_2(J, :), normals_2(J, :), areas_2(J), radiuses_2(J), &
             quad_points(J, :, :), quad_weights(J, :),                    &
             wavenumber, depth,                                           &
-            tabulated_r_range, tabulated_z_range, tabulated_integrals,   &
+            tabulation_method, tabulated_r_range, tabulated_z_range, tabulated_integrals,   &
             NEXP, AMBDA, AR,                                             &
             SP2, VSP2_SYM, VSP2_ANTISYM                                  &
           )
@@ -257,7 +258,7 @@ CONTAINS
             centers_2(J, :), normals_2(J, :), areas_2(J), radiuses_2(J), &
             quad_points(J, :, :), quad_weights(J, :),                    &
             wavenumber, depth,                                           &
-            tabulated_r_range, tabulated_z_range, tabulated_integrals,   &
+            tabulation_method, tabulated_r_range, tabulated_z_range, tabulated_integrals,   &
             NEXP, AMBDA, AR,                                             &
             SP2, VSP2_SYM, VSP2_ANTISYM                                  &
           )